Rename files to indicate existing derived target ref config is deprecated
